Ok so the difference of Tanh output is caused by tanh and tanhf, the former is for double while the latter is for float

+++ test.c
@@ -204,7 +204,7 @@ static void Tanh(
     for (p = input, q = output;
          p < (input + size) && q < (output + size);
          p++, q++) {
-        *q = tanh(*p);
+        *q = tanhf(*p);
     }
     printf("tanh: first 5 outputs: ");
     for (int i = 0; i < 5; i++)
